Refactoring flowchart nodes definition - Merged DataTypeUtils.ts to FileNodeFactory.ts
